requirements.txt
chemspipy==2.0.0
discord.py==1.5.1
fuzzywuzzy==0.17.0
cirpy==1.0.2
asyncpg==0.20.1
jishaku==1.20.0.220